clients/python3/clientsync.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
#
# Copyright (C) 2019 KuraLabs S.R.L
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.

from sys import argv
from websocket import WebSocket


def follow(filename):
    url = 'ws://localhost:9292/follow/{}'.format(filename)
    ws = WebSocket()

    try:
        ws.connect(url)
        while True:
            msg = ws.recv()
            print('{} :: {}'.format(filename, msg), end='')

    finally:
        # ws.close() takes forever
        ws.shutdown()


follow(argv[1])
