clients/python3/requirements.txt
# Async client
aiohttp

# Sync client
websocket_client
